'sat_id' now checks both the position _and_ motion of possible matches between MPC astrometry and satellites from TLEs.  As a result, it's much better both at figuring out which satellite you've see and eliminating ones that happened to be in the right area,  but moving the wrong way.  It also sped the code up a bit.
The new method is as follows :

-- Input astrometry is sorted by ID,  then by date.  We then look
for a pair of observations far enough apart to determine the object
motion accurately,  and throw out other observations.

-- Then,  for each object (now defined by a single pair of
observations),  we look for satellites within the search range of
the RA/decs in question.  Among those,  we look for those moving
within the motion tolerance range of the motion defined by that
pair of observations.  If it matches in position and motion,  it's
almost certainly a good ID,  and if it doesn't match in one or the
other,  it almost certainly isn't a good ID.